Use `@main` attribute for Lexgen entry point
